+++ file/downloadfolder_test.go
@@ -33,7 +33,7 @@ func (m *MockerDownloader) Download(_ context.Context, p files_sdk.FileDownloadP
 }
 
 type ReporterCall struct {
-	status.Report
+	status.File
 	err error
 }
 
@@ -53,9 +53,9 @@ func NewTestSetup() *TestSetup {
 	return t
 }
 
-func (setup *TestSetup) Reporter() Reporter {
-	return func(status status.Report, err error) {
-		setup.reporterCalls = append(setup.reporterCalls, ReporterCall{Report: status, err: err})
+func (setup *TestSetup) Reporter() status.Reporter {
+	return func(status status.File, err error) {
+		setup.reporterCalls = append(setup.reporterCalls, ReporterCall{File: status, err: err})
 	}
 }
 
@@ -70,7 +70,7 @@ func (setup *TestSetup) TearDown() error {
 	return os.RemoveAll(setup.tempDir)
 }
 
-func (setup *TestSetup) Call() (status.Job, error) {
+func (setup *TestSetup) Call() status.Job {
 	return downloadFolder(
 		context.Background(),
 		setup.files,
@@ -93,25 +93,24 @@ func Test_downloadFolder_ending_in_slash(t *testing.T) {
 	setup.files = append(setup.files, Entity{file: files_sdk.File{Path: "some-path/taco.png", Size: 100, Type: "file"}})
 	setup.DownloadFolderParams = DownloadFolderParams{FolderListForParams: files_sdk.FolderListForParams{Path: "some-path"}, Reporter: setup.Reporter(), RootDestination: setup.RootDestination()}
 	setup.rootDestination = "some-path/"
-	_, err := setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
-	assert.Equal(1, setup.reporterCalls[0].Job().Count())
+	assert.Equal(1, setup.reporterCalls[0].Job.Count())
 	assert.Equal(4, len(setup.reporterCalls))
-	assert.Equal(status.Queued, setup.reporterCalls[0].Type())
-	assert.Equal(status.Downloading, setup.reporterCalls[1].Type())
-	assert.Equal(status.Downloading, setup.reporterCalls[2].Type())
-	assert.Equal(status.Complete, setup.reporterCalls[3].Type())
-	assert.Equal("some-path/taco.png", setup.reporterCalls[0].File().Path)
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
+	assert.Equal(status.Queued, setup.reporterCalls[0].Status)
+	assert.Equal(status.Downloading, setup.reporterCalls[1].Status)
+	assert.Equal(status.Downloading, setup.reporterCalls[2].Status)
+	assert.Equal(status.Complete, setup.reporterCalls[3].Status)
+	assert.Equal("some-path/taco.png", setup.reporterCalls[0].File.Path)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
 	assert.Equal(1, len(setup.downloader.Calls))
 	assert.Equal("some-path/taco.png", setup.downloader.Calls[0].Path)
 
-	assert.Equal(true, setup.reporterCalls[0].Job().AllEnded())
-	assert.Equal(int64(100), setup.reporterCalls[0].Job().TransferBytes())
-	assert.Equal(int64(100), setup.reporterCalls[0].Job().TotalBytes())
+	assert.Equal(true, setup.reporterCalls[0].Job.AllEnded())
+	assert.Equal(int64(100), setup.reporterCalls[0].Job.TransferBytes())
+	assert.Equal(int64(100), setup.reporterCalls[0].Job.TotalBytes())
 
 	assert.NoError(setup.TearDown())
 }
@@ -124,19 +123,18 @@ func Test_downloadFolder_more_than_one_file(t *testing.T) {
 	setup.DownloadFolderParams = DownloadFolderParams{FolderListForParams: files_sdk.FolderListForParams{Path: "some-path"}, Reporter: setup.Reporter(), RootDestination: setup.RootDestination()}
 	setup.rootDestination = "some-path"
 
-	_, err := setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
-	assert.Equal(2, setup.reporterCalls[0].Job().Count())
+	assert.Equal(2, setup.reporterCalls[0].Job.Count())
 	assert.Equal(8, len(setup.reporterCalls))
-	assert.ElementsMatch([]string{"some-path/taco.png", "some-path/pizza.png"}, []string{setup.reporterCalls[0].File().Path, setup.reporterCalls[1].File().Path})
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
+	assert.ElementsMatch([]string{"some-path/taco.png", "some-path/pizza.png"}, []string{setup.reporterCalls[0].File.Path, setup.reporterCalls[1].File.Path})
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
 	assert.Equal(2, len(setup.downloader.Calls))
 	assert.ElementsMatch([]string{"some-path/taco.png", "some-path/pizza.png"}, []string{setup.downloader.Calls[0].Path, setup.downloader.Calls[1].Path})
 
-	assert.Equal(true, setup.reporterCalls[1].Job().AllEnded())
-	assert.Equal(int64(200), setup.reporterCalls[1].Job().TransferBytes())
-	assert.Equal(int64(200), setup.reporterCalls[1].Job().TotalBytes())
+	assert.Equal(true, setup.reporterCalls[1].Job.AllEnded())
+	assert.Equal(int64(200), setup.reporterCalls[1].Job.TransferBytes())
+	assert.Equal(int64(200), setup.reporterCalls[1].Job.TotalBytes())
 
 	assert.NoError(setup.TearDown())
 }
@@ -149,13 +147,12 @@ func Test_downloadFolder_sync_already_downloaded(t *testing.T) {
 	setup.rootDestination = ""
 	_, err := os.Create(filepath.Join(setup.tempDir, "taco.png"))
 	assert.NoError(err)
-	_, err = setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
 	assert.Equal(2, len(setup.reporterCalls))
-	assert.Equal(status.Queued, setup.reporterCalls[0].Type())
-	assert.Equal(status.Skipped, setup.reporterCalls[1].Type())
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
+	assert.Equal(status.Queued, setup.reporterCalls[0].Status)
+	assert.Equal(status.Skipped, setup.reporterCalls[1].Status)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
 	assert.Equal(0, len(setup.downloader.Calls))
 
 	assert.NoError(setup.TearDown())
@@ -169,16 +166,15 @@ func Test_downloadFolder_sync_not_already_downloaded(t *testing.T) {
 	setup.rootDestination = ""
 	_, err := os.Create(filepath.Join(setup.tempDir, "taco.png"))
 	assert.NoError(err)
-	_, err = setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
 	assert.Equal(4, len(setup.reporterCalls))
-	assert.Equal("taco.png", setup.reporterCalls[0].File().Path)
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
-	assert.Equal(status.Queued, setup.reporterCalls[0].Type())
-	assert.Equal(status.Downloading, setup.reporterCalls[1].Type())
-	assert.Equal(status.Downloading, setup.reporterCalls[2].Type())
-	assert.Equal(status.Complete, setup.reporterCalls[3].Type())
+	assert.Equal("taco.png", setup.reporterCalls[0].File.Path)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
+	assert.Equal(status.Queued, setup.reporterCalls[0].Status)
+	assert.Equal(status.Downloading, setup.reporterCalls[1].Status)
+	assert.Equal(status.Downloading, setup.reporterCalls[2].Status)
+	assert.Equal(status.Complete, setup.reporterCalls[3].Status)
 	assert.Equal(1, len(setup.downloader.Calls))
 	assert.Equal("taco.png", setup.downloader.Calls[0].Path)
 
@@ -191,10 +187,10 @@ func Test_downloadFolder_Entity_error(t *testing.T) {
 	setup.files = append(setup.files, Entity{file: files_sdk.File{Path: "taco.png", Size: 100, Type: "file"}, error: fmt.Errorf("something Happened")})
 	setup.DownloadFolderParams = DownloadFolderParams{FolderListForParams: files_sdk.FolderListForParams{Path: "some-path"}, Reporter: setup.Reporter(), RootDestination: setup.RootDestination()}
 	setup.rootDestination = "some-path/"
-	job, err := setup.Call()
-	assert.Equal(0, len(setup.reporterCalls))
-	assert.Errorf(err, "something Happened")
-	assert.Equal(0, job.Count())
+	job := setup.Call()
+	assert.Equal(1, len(setup.reporterCalls))
+	assert.Errorf(setup.reporterCalls[0].err, "something Happened")
+	assert.Equal(1, job.Count())
 	assert.Equal(true, job.AllEnded())
 	assert.Equal(int64(0), job.TotalBytes())
 	assert.Equal(int64(0), job.TransferBytes())
@@ -209,13 +205,12 @@ func Test_downloadFolder_download_file(t *testing.T) {
 	setup.DownloadFolderParams = DownloadFolderParams{FolderListForParams: files_sdk.FolderListForParams{Path: "some-path"}, Reporter: setup.Reporter(), RootDestination: setup.RootDestination()}
 	setup.rootDestination = "taco.png"
 
-	job, err := setup.Call()
-	assert.NoError(err)
+	job := setup.Call()
 
 	assert.Equal(1, job.Count())
 	assert.Equal(4, len(setup.reporterCalls))
-	assert.Equal("taco.png", setup.reporterCalls[0].File().Path)
-	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes())
+	assert.Equal("taco.png", setup.reporterCalls[0].File.Path)
+	assert.Equal(int64(0), setup.reporterCalls[0].TransferBytes)
 	assert.Equal(1, len(setup.downloader.Calls))
 	assert.Equal("taco.png", setup.downloader.Calls[0].Path)
 
@@ -229,13 +224,12 @@ func Test_downloadFolder_OnDownload(t *testing.T) {
 	setup.DownloadFolderParams = DownloadFolderParams{FolderListForParams: files_sdk.FolderListForParams{Path: "some-path"}, Reporter: setup.Reporter(), RootDestination: setup.RootDestination()}
 	setup.rootDestination = "taco.png"
 
-	_, err := setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
 	assert.Equal(4, len(setup.reporterCalls))
 	assert.Equal(1, len(setup.downloader.Calls))
 
-	assert.Equal(int64(100), setup.reporterCalls[3].File().Size, "Updates with real file size")
+	assert.Equal(int64(100), setup.reporterCalls[3].File.Size, "Updates with real file size")
 
 	assert.NoError(setup.TearDown())
 }
@@ -248,12 +242,11 @@ func Test_downloadFolder_Download_error(t *testing.T) {
 	setup.rootDestination = "taco.png"
 	setup.downloader.ReturnError = fmt.Errorf("download error")
 
-	_, err := setup.Call()
-	assert.NoError(err)
+	setup.Call()
 
 	assert.Equal(4, len(setup.reporterCalls))
-	assert.Equal("taco.png", setup.reporterCalls[0].File().Path)
-	assert.Contains(setup.reporterCalls[0].Destination(), "taco.png")
+	assert.Equal("taco.png", setup.reporterCalls[0].File.Path)
+	assert.Contains(setup.reporterCalls[0].LocalPath, "taco.png")
 	var errStatus error
 	for _, r := range setup.reporterCalls {
 		if r.err != nil {
@@ -263,4 +256,27 @@ func Test_downloadFolder_Download_error(t *testing.T) {
 	assert.Errorf(errStatus, "download error")
 
 	assert.NoError(setup.TearDown())
+}
+
+func Test_tmpDownloadPath(t *testing.T) {
+	assert := assert.New(t)
+
+	path := tmpDownloadPath("you-wont-find-me")
+
+	assert.Equal("you-wont-find-me.download", path)
+	file, err := os.Create("find-me.download")
+	defer func() {
+		os.Remove(file.Name())
+	}()
+	if err != nil {
+		panic(err)
+	}
+	file.Write([]byte("hello"))
+	err = file.Close()
+	if err != nil {
+		panic(err)
+	}
+	path = tmpDownloadPath("find-me")
+
+	assert.Equal(fmt.Sprintf("%v (1)", file.Name()), path, "it increments a number")
 }