+++ file/status/job.go
@@ -1,32 +1,49 @@
 package status
 
 import (
+	"context"
 	"sync"
 
-	filesSDK "github.com/Files-com/files-sdk-go"
 	"github.com/chilts/sid"
+
+	filesSDK "github.com/Files-com/files-sdk-go"
 )
 
-type Job struct {
-	statuses      []Report
-	statusesMutex *sync.Mutex
-	ended         int
-	Id            string
+type ToStatusFile interface {
+	ToStatusFile() File
 }
 
-func (r Job) Init() *Job {
-	r.statusesMutex = &sync.Mutex{}
+type Job struct {
+	statuses      []ToStatusFile
+	statusesMutex *sync.RWMutex
+	Id            string
+	context.CancelFunc
+}
+
+func (r Job) Init(id ...string) *Job {
+	r.statusesMutex = &sync.RWMutex{}
+	if len(id) == 1 {
+		r.Id = id[0]
+	}
 	if r.Id == "" {
 		r.Id = sid.IdBase64()
 	}
 	return &r
 }
 
+func (r Job) Cancel() {
+	r.CancelFunc()
+}
+
+func (r Job) Job() Job {
+	return r
+}
+
 func (r Job) Count() int {
 	return len(r.statuses)
 }
 
-func (r *Job) Add(report Report) {
+func (r *Job) Add(report ToStatusFile) {
 	r.statusesMutex.Lock()
 	r.statuses = append(r.statuses, report)
 	r.statusesMutex.Unlock()
@@ -34,46 +51,47 @@ func (r *Job) Add(report Report) {
 
 func (r Job) TotalBytes() int64 {
 	var total int64
-	r.statusesMutex.Lock()
+	r.statusesMutex.RLock()
 	for _, s := range r.statuses {
-		if s.Valid() {
-			total += s.File().Size
+		if s.ToStatusFile().Valid() {
+			total += s.ToStatusFile().Size
 		}
 	}
-	r.statusesMutex.Unlock()
+	r.statusesMutex.RUnlock()
 	return total
 }
 
 func (r Job) TransferBytes() int64 {
 	var transferBytes int64
-	r.statusesMutex.Lock()
+	r.statusesMutex.RLock()
 	for _, s := range r.statuses {
-		if s.Valid() {
-			transferBytes += s.TransferBytes()
+		if s.ToStatusFile().Valid() {
+			transferBytes += s.ToStatusFile().TransferBytes
 		}
 	}
-	r.statusesMutex.Unlock()
+	r.statusesMutex.RUnlock()
 	return transferBytes
 }
 
 func (r Job) AllEnded() bool {
 	allEnded := true
-	r.statusesMutex.Lock()
+	r.statusesMutex.RLock()
 	for _, s := range r.statuses {
-		if !s.Ended() {
+		if !s.ToStatusFile().Ended() {
 			allEnded = false
+			break
 		}
 	}
-	r.statusesMutex.Unlock()
+	r.statusesMutex.RUnlock()
 	return allEnded
 }
 
 func (r Job) Files() []filesSDK.File {
 	var files []filesSDK.File
-	r.statusesMutex.Lock()
+	r.statusesMutex.RLock()
 	for _, s := range r.statuses {
-		files = append(files, s.File())
+		files = append(files, s.ToStatusFile().File)
 	}
-	r.statusesMutex.Unlock()
+	r.statusesMutex.RUnlock()
 	return files
 }