+++ file/uploadstatus.go
@@ -9,13 +9,28 @@ import (
 )
 
 type UploadStatus struct {
-	file files_sdk.File
+	files_sdk.File
 	status.Status
-	destination   string
-	Source        string
-	cancel        context.CancelFunc
+	context.CancelFunc
+	Job           *status.Job
+	LocalPath     string
+	RemotePath    string
 	UploadedBytes int64
-	job           *status.Job
+	Sync          bool
+}
+
+func (u UploadStatus) ToStatusFile() status.File {
+	return status.File{
+		TransferBytes: u.UploadedBytes,
+		File:          u.File,
+		Cancel:        u.CancelFunc,
+		LocalPath:     u.LocalPath,
+		RemotePath:    u.RemotePath,
+		Id:            u.Id(),
+		Job:           u.Job,
+		Status:        u.Status,
+		Sync:          u.Sync,
+	}
 }
 
 func (u *UploadStatus) SetStatus(s status.Status) {
@@ -26,31 +41,8 @@ func (u *UploadStatus) SetStatus(s status.Status) {
 	u.Status = s
 }
 
-func (u UploadStatus) Cancel() {
-	if u.cancel != nil {
-		u.cancel()
-		u.Status = status.Canceled
-	}
-}
-
-func (u UploadStatus) TransferBytes() int64 {
-	return u.UploadedBytes
-}
-
-func (u UploadStatus) File() files_sdk.File {
-	return u.file
-}
-
-func (u UploadStatus) Destination() string {
-	return u.destination
-}
-
-func (u UploadStatus) Job() status.Job {
-	return *u.job
-}
-
 func (u UploadStatus) Id() string {
-	return u.Job().Id + ":" + u.File().Path
+	return u.Job.Id + ":" + u.File.Path
 }
 
 func (u *UploadStatus) incrementDownloadedBytes(b int64) {